+++ test/spec/manager.state.ut.js
@@ -17,7 +17,9 @@
             beforeEach(function() {
                 currentUser = {
                     id: 'u-1',
-                    org: 'o-fn8y54thf85',
+                    org: {
+                        id: 'o-fn8y54thf85'
+                    },
                     username: 'test'
                 };
 
@@ -83,7 +85,7 @@
                 });
 
                 it('should get all the minireels that are associated with the user\'s org', function() {
-                    expect(cinema6.db.findAll).toHaveBeenCalledWith('experience', { type: 'minireel', org: currentUser.org, sort: 'lastUpdated,-1' });
+                    expect(cinema6.db.findAll).toHaveBeenCalledWith('experience', { type: 'minireel', org: currentUser.org.id, sort: 'lastUpdated,-1' });
                 });
 
                 it('should return the model it already has if it has one', function() {